reduce duplicate code
